ai_novel/backend/requirements.txt
Django>=4.2.7,<5.0.0
djangorestframework>=3.14.0,<4.0.0
psycopg2>=2.9.6,<3.0.0
requests>=2.31.0,<3.0.0
celery>=5.3.1,<6.0.0
redis>=4.6.0,<5.0.0
django-cors-headers>=4.1.0,<5.0.0
django-filter>=23.2,<24.0
dj-rest-auth>=4.0.1,<5.0.0
django-environ>=0.11.2,<1.0.0
gunicorn>=21.2.0,<22.0.0
whitenoise>=6.5.0,<7.0.0
python-json-logger>=2.0.7,<3.0.0
drf-spectacular>=0.26.4,<1.0.0
